dnekcab-eciovni/migracao/requirements.txt
jinja2==2.10
mysql-connector==2.1.6
